_pluginfactory/pluginfactory.py: Implement error reporting for pip origins
For plugins loaded from the pip origin, we now support specifying constraints.

This is a non-breaking change and only involves specifying a package
with constraints when specifying the 'package-name', however there are
a few errors errors which can occur as a result, this patch tries to
handle them all cleanly and provide the user with useful error messages.