_loader/loadelement.pyx: Load full element names recursively.
When opening up the possibility of loading the same junction twice,
it becomes possible to have multiple elements with the same name
following the current implementation logic which enforces that
you must have a toplevel junction to coalesce junctions and synchronize
their configuration.

Since we no longer coalesce junction names, we need to allow elements
to be reported in the UI with their full depth.