_ostree.py: Fix call to remote_gpg_import() again
Last time we fixed this, we actually only adjusted it to work with
the new API but stopped supporting the old API, which is still widely
in use (debian 9, debian 10 and ubuntu 18 all still use ostree v2019.1).

This time, lets add an except block here and fallback to the older
API if the new API raises a TypeError.