Allow junctions to inject dependencies from current project to junctioned one
There are cases where downstream projects want to overwrite upstream
elements.

So far, either the downstream project just allows overlaps and hopes
that ABI is compatible. There are also files left from original
elements when not overwritten.

Or downstream adds an "overlay" with local source to the
junctions. However on the latter, elements in the overlay are
considered as part of upstream, which means in cannot depend on
element outside of upstream (unless cyclic junction maybe?).

Here we add a feature to junctions to allow injecting elements from
current project to junctions.

This is useful for example in the case of GNOME SDK needing to override
GLib, GObjectIntrospection, libsoup, etc. from Freedesktop SDK.