_scheduler: Fix order of launching jobs and sending notifications.
Sending notifications causes potentially large bodies of code to run
in the abstracted frontend codebase, we are not allowed to have knowledge
of the frontend from this code.

Previously, we were adding the job to the active jobs, sending the
notification, and then starting the job. This means that if a BuildStream
frontend implementation crashes, we handle the excepting in an inconsistent
state and try to kill jobs which are not running.

In addition to making sure that active_jobs list adjustment and
job starting does not have any code body run in the danger window
in between these, this patch also adds some fault tolerance and assertions
around job termination so that:

  o Job.terminate() and Job.kill() do not crash with None _process

  o Job.start() raises an assertion if started after being terminated

This fixes the infinite looping aspects of frontend crashes at
job_start() time described in #1312.