junctions: Renaming `target` option to `link`.
The word `target` is confusing as by itself, it implies the opposite
of what the option actually does. Without additional context,
the word `target` seems to indicate "The target junction declaration
in a subproject to overwrite with this junction declaration", when
in fact we are inheriting the configuration rather than squashing it.

After discussing this, it appears the word `target` was selected
as it matches with filesystem terminology, e.g. the target specified
to the `ln` command when creating symbolic or hard links.

As such, this patch renames `target` to `link` instead, as this
should be less confusing, it indicates that we are creating a link
of two junctions; the parameter of the `link` can then be loosely
referred to as a `target`, but at least the word `link` is used.

Changes:

  * junction.py:
    - Rename of `target` to `link`, updating relevant documentation.
    - Enhanced documentation to link to the main dependencies section
      and use a less dated example of depending on an element through
      a junction
    - Updated preflight() method to:
      - Be a little more simplified (less branching)
      - Provide machine readable errors
      - Provide new errors in case other options are provided along
        with the link option, as these will be ignored.

  * _loader/loader.py: Updated to use the new `link` junction members instead of `target`

  * tests/format/junctions.py: Updated tests to:
    - Use new `link` terminology
    - Use machine readable errors instead of asserting strings from the UI
    - Improve coverage of junction misconfigurations, grouping the
      testing code under a single parameterized test