sandbox/_sandboxremote.py: Improved log messages
The log message which shows that the remote build is ongoing has been
extended to include the name of the element to which the remote build
corresponds.

Likewise the message indicating that an action result has been found in
the action cache will also no include the element name.

A new timed_activity() call has been added to show when missing blobs
are being determined and uploaded.

Fixed #1336