_loader: Fix invalid junction error to include provenance
When a junction element has dependencies, we were not even specifying
which junction element was guilty of having dependencies and only
saying that such a thing would be forbidden (wherever it might be in
your project).

Fix this by using the provenance of the dependency in the junction,
specifying exactly where the invalid dependency exists.

This required carrying the provenance onto the other Dependency() object.