_yaml.pdx: Remove false presumption from EISDIR error message
Change the error message to reflect that the file is a directory,
do not suggest that it should be a .bst file, it should not be
a .bst file.

A .bst file might be expected from the loader code, which can
feel free to express that, but other files can be loaded, for
example include files.