_includes.py: Propagate provenance through Loader.get_loader()
Instead of raising a customized error message which adds little
value to the provenance, just pass the provenance along.

This is important so that the Loader is aware of the provenance
of loaded junctions, so that it can more precisely report errors
about conflicting junctions when includes cause conflicts.

This commit also adjusts tests/format/includes.py