_frontent/cli.py: Removed message to user informing the workspace is closed.
When closing a workspace while the `workspace close` command was issued
from the closed workspace, we go to great lengths to inform the user
that they can no longer run bst commands from the closed workspace directory.

This is not worthwhile, the user will know this the next time they run
`bst` anyway.