_stream.py: Remove workspace_is_required()
This API was only used to issue a not very useful message in the CLI
when removing a workspace the user happened to have invoked BuildStream
from.

The API also didn't really make any sense, being documented as:

  "Checks whether the workspace belonging to element_name is
   required to load the project"

This in fact meant whether having that workspace open was required
for BuildStream to be invoked, due to the metadata encoded into the
workspace directory.