tests/frontend: Re-introduce tests for `track --except`
Previously, the `--except` functionality was only tested as part of
`build --track`. And, accidentally got removed while support for
`--track` was being ripped out from `bst --build`.

Add a new and improved version of that test, using `source track`
command instead, and also exercising the `--deps` option in conjunction
with `--except`. In future, if we support more values for `--deps`
(like `build` and `run`), it should be relatively easy to extend the
same test.

This was originally brought up in
https://gitlab.com/BuildStream/buildstream/-/merge_requests/1893#note_335696186.