_variables.pyx: Setting up code for split of fast/slow paths
Use _expand_var() and _expand_value_expression() as gateways for
handling KeyError and RecursionError and still calling _check_variables()
for now as a fallback.

Moved implementations to _fast_expand_var() and _fast_expand_value_expression()
and in a later commit we will fallback to the non-recursive algorithm, and
remove _check_variables() completely as this will be obsoleted.