_variables.py: Complete rewrite
Rebasing bst-1 variables code on bst-2 cython version, without the cython.

This is mostly a drop in replacement, except that:

  * Now calls to Variables.subst() must provide the passed string's Provenance

  * Elements must now call Variables.check() explicitly

This commit also adds a test to ensure that junctions can use partially
resolved variables, even when the calling project would require that all
variables be resolved in order to completely be resolved. Resolving variables
within junctions is allowed so long as those variable resolutions do not
require variable values from the target project they are junctioning.