_yaml.py: Support merging of conditional lists
During composition of project.conf, it can happen that we
are including and composing node hierarchies before options
are resolved and before we are ready to evaluate conditional
statements.

Without this patch, conditional statements are treated like
regular lists, causing existing conditional statements to
be overwritten by subsequently composited conditional statements
without ever having been resolved.

This patch introduces a special case for composing lists,
when composing a sequence that is a conditional statement, the
sequence is appended instead of overwriting the underlying
list.

Note that precedence of declarations in including vs included
yaml fragments did raise some concern, and a test in the following
commit is added to cover this case. Composition of conditional lists
on top of other conditional lists are unconditionally appended,
however the semantics of includes ensures that the including fragment's
conditionals is always composited *on top* over the included fragment,
ensuring that we have the correct pecedence even when compositing
yet to be resolved conditional directives.