.pylintrc: Simplify config, only list overrides
Rather than having the full `.pylintrc`, only list the overrides that we
have compared to the default options. This makes it easy to identify
what options we have modified, which is also similar to what we do for
configuring other tools.

At the same time, also remove overrides corresponding to the `gi`
module. They were added when this repository depended on the `ostree`
module but that's no longer the case.