_project.py: Only one set of plugin factories
This commit makes it mandatory to specify the `plugins` configuration
directly in the `project.conf`, and at the same time removes any ambiguity
as to which plugins are loaded and how. It thus becomes impossible for
the first pass to load the same plugins differently than in the second
pass.

Additionally, some additional project.conf keys are asserted to be
specified in the toplevel project.conf and not included as a matter
of policy and consistency.

Summary of changes:

  * _project.py

    - Only one set of plugin factories, loaded only in the first pass.
    - Assert the following keys are only specified in project.conf and
      not included from other files:
        name, element-path, min-version, plugins
    - The create_element() and create_source() methods no longer need
      to take a `first_pass` argument

  * element.py: Updated to not specify `first_pass` to create_element()
    or create_source()

  * _frontend/widget.py: Print the only set of plugins which are loaded,
    no more ambiguity about first or second pass.

  * _pluginfactory/pluginoriginjunction.py: Chain load plugins from
    the single element/source factory, instead of accessing the
    Project.config member which no longer has the factories.