_yaml.pyx: Make shortname a mandatory argument to _yaml.load()
This ensures that important calls to this function do give some
thought to providing a reasonable shortname, which will be used
as a display name in errors.

This continues to support `None` as a shortname, which is used
in various tests which don't need to provide a reasonable user
facing error.

The buildstream.testing module now exports a `load_yaml` function
which only takes a filename and no shortname.