_loader: Use only one Dependency() class
This moves the Dependency() implementation from the _loader.types
module into the _loader.loadelement module, replacing the duplicate
Dependency() definition there.

Instead of creating _loader.loadelement.Dependency objects on the
fly during the load process, we simply fill in the new LoadElement
member on the original Dependency objects.

This refactor should make the code easier to work with, since any
originally loaded state from the YAML is preserved in the intermediate
LoadElement data model and does not need to be manually handed over
to a separate object.

Summary of changes:

  * _loader/types.pyx: Removed Dependency() implementation from here,
    and renamed to _loader/types.py

  * _loader/loadelement.py: Replaced the second Dependency object
    implementation with the original one from _loader/types.py,
    adding a new LoadElement member which is resolved during the
    load process via the new Dependency.set_element() API

  * _loader/loader.py: Instead of creating a second kind of Dependency
    object on the fly during the load process, simply resolve the
    existing dependencies with Dependency.set_element()

  * setup.py: Fixed up cython build instructions, now that _loader/types.py
    no longer contains any cython code.