_elementsources.py: Fix source names in unique key
The goal was to include the source plugin kind in the element cache key
as the unique key of a source may not be unique across different source
plugins. This is the source equivalent of the `element-plugin-name`
value in the element cache key.

However, `Source._get_source_name()` was the wrong method for this as
that also includes the key itself, which may not even be set yet.

This results in a cache key change.

Fixes: 3953bcc6 ("element.py: clobber sources with workspace")