Move handling of the source `directory` configuration to ElementSources
The `directory` value determines where a source is staged within the
build root of an element, however, it does not directly affect
individual sources.

With this change the sources will individually be cached in CAS
independent of the value of `directory`. `ElementSources` will use the
value of `directory` when staging all element sources into the build
root.

This results in a cache key change as the `directory` value is moved
from the unique key of individual sources to the unique key of
`ElementSources`.

This is in preparation for #1274.