buildstream/element.py: Consider "build-root" variable in the cache key unconditionally
Not all elements use the "build-root" variable, but it is the standard
variable to use for the build directory, and the build directory must
be considered in the cache key.

Handling this unilaterally in the core is safer than delegating this
to element implementations, as we have less chance of plugin authors
missing this detail and possibly introducing binary variance for
artifacts where only the build directory differs (something which
happens when the project name or element names change).

This commit also updates the hard coded cache keys in the cache key
test, so as to ensure every commit passes it's own tests.

This fixes #1386.