utils.py: Replace the (?ms) switch by equivalent re. options
The 'm' switch stands for multiline output.
The 's' makes '.' also match newlines.

It is problematic in mutliple places where its use is deprecated,
as now they need to be at the start of an expression, which is not the
case in all the calls to it that we are doing.

If someone wants the same functionality, they should add it explictely
after calling _glob2re