utils.py: Remove the (?ms) switch at the start of _glob2re
The ?ms switch stands for multiline output.

This switch is used nowhere in our codebase.
It also is problematic in mutliple places where its use is deprecated,
as now they need to be at the start of an expression, which is not the
case in all the calls to it that we are doing.

If someone wants the same functionality, they should add it explictely
after calling _glob2re