scriptelement.py: Remove legacy manual creation of staging directory.
This branch statement comes from a time before virtual directories
when we were using os.makedirs() to create the target directory, and
was then ported to use virtual directories for that purpose.

However, during the same migration, the Element.stage_artifact()
function was made to guarantee the existence of the staging root
location, with a call to:

    vbasedir.descend( ... staging directory ..., create=True)

This makes the branch statement in scriptelement.py redundant.