_loader/loadelement.pyx: Merge duplicate dependencies.
When the same element is specified multiple times as a direct dependency,
merge and accumulate the results into the already loaded dependency.

  * If the same element is a runtime and build dependency separately, it
    will be a single dependency of both runtime and build.

  * If either of the dependencies are `strict`, it will be a strict
    dependency.

The build graph retains the invariant that an element only ever depends
on another element once directly, only the YAML can express the same
dependency differently more than once, and the results are accumulated.

This consequently remoces LoadErrorReason.DUPLICATE_DEPENDENCY as this is
no longer relevant.