element.py: Stop computing cache keys when caching, we should have them
This removes a call to `__update_cache_key_non_strict()` in the job for
non strict builds. This call has an impact on other elements and should
never be done beforehand.

This however means we need to compute cache keys before, and we
therefore need to re-compute the cache keys as an element become
buildable