Update requirements for tests
This commit is intended to be purely an update of the frozen
requirements. However we'd get new warnings with the new version of
pylint. For that reason, the pylint requirement has now been pinned to
what we had in the frozen requirements before.

Each time we want to upgrade pylint, we will need to change this
manually.

---

Among other things, I have noticed that older versions of grpcio take it
much longer (comparatively) to be installed on Python 3.9 compared to
the newer versions.