_assetcache.py: Allow explicit re-initialization of remotes.
This is merely because the structure of remotes is heavily bound to
loaded projects, and in some cases we need to download some artifacts
before some ArtifactProjects can exist and then later pull elements
related to those newly created ArtifactProject instances.

This is rather a silly workaround, instead the data structure of
remotes should be reorganized, which is a larger task and probably appropriate
to examine when re-examining remote service configuration surfaces.