element.py: Added internal API _mimic_artifact()
This function causes the element to mimic it's artifact instance,
and is initially used when instantiating from an artifact.

This needs to be exposed so that ArtifactElement can trigger this
activity in response to _pull_done().

Additionally, this adds the declaration of __environment and
__variables to the __init__ function, as I received some complaints
from pylint and mypy from this commit.