tests/internals/pluginloading.py: Removed test
This internal test was relying on internal construction of pipeline, and
relying on the shape of internal API interactions, which is the case for
many tests in the tests/internals directory, and thus they make internal
API churn harder to deal with.

In this case, this test has been long superceded by the newer test in
tests/plugins/loading.py, which not only tests the ability to load a custom
plugin and source successfully, but also tests the various possible failure
modes.

In short, this internal test is now thankfully obsolete.