_types.pyx: Fixing new error from updated sphinx
Our docs build does not specify a specific version of sphinx, and in
a recent update of sphinx, it will try to mess around with docstring
annotations on loaded python objects.

Our MetaFastEnum class raises a ValueError when attempting to dynamically
set a new attribute on the FastEnum, sphinx is happy to ignore an
AttributeError in this case, but has the opinion that a ValueError is
not a valid error for this problem, and thus ignores ValueError, causing
the docs build to break.

We don't really care if it is a ValueError or AttributeError that happens
in this case, we rather only care that it be raised as a BUG at runtime
if a value is assigned dynamically, so lets raise an AttributeError instead
and satisfy sphinx's opinion.