ci: Simplify pipelines using Docker Compose
Replace our bespoke and rather long to type `docker` CLI invocations
with more manageable docker-compose commands. This should also be useful
for better replicating what the CI does on laptops, and also to help
users to run tests inside a container (especially when that's not
possible natively like on MacOS).

On a technical note, Compose does support YAML anchors, unlike GitHub
Actions. That allows us to reuse the common configuration options
without having to duplicate them everywhere.

This also removes most of the duplication in the image names.