source.py: Raise a fatal-able warning when using an unaliased source.
We do this at Source.mark_download_url() time, which as a matter of policy,
must be used at project configure() time if Source.translate_url() is not
used instead, this should ensure that aliases are verified to be used
at load time, if not, it is a bug in the respective plugin.

We will not be able to catch unaliased URLs unless they are specified
in the yaml, for example, if you have not specified all of the submodules
in use for a git source.