tests/plugins/loading.py: Refactor test to not depend on yaml.dump() formatting
This test was writing the YAML with ruamel.yaml and then asserting that we
have the correct provenance, however as ruamel.yaml evolves we cannot be sure
that the dump formatting will be exactly the same (as it has in this instance).

To avoid confusion, create static copies of the project.conf so that the
provenance can be tested properly.