.github/: Remove the 'no-usedevelop' CI
This job used to run standard CI with the `nocover` flag to ensure that
the tox env would not set things up with the `usedevelop` flag, which has
historically been required for coverage collection for some reason.

At this point, we run everything post py37 with the `nocover` flag because
all these python versions have broken the cython/coverage combination.

Since we *mostly* run CI with `nocover`, there is no point to keep this
job around for now.